Migrated ConsoleMeDataTable to React Hooks
