+++ ui/src/components/request/RequestTable.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { Header } from "semantic-ui-react";
-import ConsoleMeDataTable from "../blocks/ConsoleMeDataTable";
+import ConsoleMeDataTable from "../blocks/datatable/DataTableComponent";
 import { useAuth } from "../../auth/AuthProviderDefault";
 import ReactMarkdown from "react-markdown";
